=。=

+++ src/main/java/com/websit/mapper/SpecificationsMapper.java
@@ -55,5 +55,5 @@ public interface SpecificationsMapper {
 	 * @author pangchong
 	 * @createDate 2019年3月23日 下午2:00
 	 */
-	public int deleteSpecificationsVoById(@Param("specificationsId") BigInteger specificationsId);
+	public int deleteSpecificationsVoById(@Param("specificationsId") Long specificationsId);
 }
